Changed the Google Login page to a Django view.
