+++ seqr/urls.py
@@ -14,6 +14,7 @@ from django.contrib import admin
 import django.contrib.admindocs.urls
 import django.views.static
 
+from seqr.views.utils.google_login import google_login_view, google_grant_view
 from seqr.views.apis.family_api import \
     update_family_fields_handler, \
     edit_families_handler, \
@@ -134,7 +135,6 @@ react_app_pages = [
 no_login_react_app_pages = [
     r'^$',
     'login',
-    'oauth2callback',
     'users/forgot_password',
     'users/set_password/(?P<user_token>.+)',
     'matchmaker/matchbox',
@@ -265,6 +265,10 @@ api_endpoints = {
 }
 
 urlpatterns = []
+
+# Google auth
+urlpatterns += [url('^google_login$', google_login_view)]
+urlpatterns += [url('^oauth2callback$', google_grant_view)]
 
 # core react page templates
 urlpatterns += [url("^%(url_endpoint)s$" % locals(), main_app) for url_endpoint in react_app_pages]
