+++ ui/app.jsx
@@ -15,7 +15,6 @@ import GeneInfoSearch from 'pages/GeneInfoSearch'
 import LocusLists from 'pages/LocusLists'
 import Staff from 'pages/Staff/Staff'
 import Login from 'pages/Login/components/Login'
-import GoogleAuth from 'pages/Login/components/GoogleAuth'
 import ForgotPassword from 'pages/Login/components/ForgotPassword'
 import SetPassword from 'pages/Login/components/SetPassword'
 import LandingPage from 'pages/Public/LandingPage'
@@ -58,7 +57,6 @@ ReactDOM.render(
             <Route path="/gene_lists" component={LocusLists} />
             <Route path="/staff" component={Staff} />
             <Route path="/login" component={Login} />
-            <Route path="/oauth2callback" component={GoogleAuth} />
             <Route path="/users/forgot_password" component={ForgotPassword} />
             <Route path="/users/set_password" component={SetPassword} />
             <Route path="/matchmaker/matchbox" component={MatchmakerInfo} />
